# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.1
paths:
...
  /api/v1/map/clusters:
    get:
      tags:
      - map
      summary: List Clusters
      operationId: list_clusters_api_v1_map_clusters_get
      parameters:
      - name: owner_faction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClusterSummary'
                title: Response List Clusters Api V1 Map Clusters Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/sectors:
    get:
      tags:
      - map
      summary: List Sectors
      operationId: list_sectors_api_v1_map_sectors_get
      parameters:
      - name: cluster_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
      - name: owner_faction
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SectorSummary'
                title: Response List Sectors Api V1 Map Sectors Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/sectors/{sector_id}:
    get:
      tags:
      - map
      summary: Get Sector
      operationId: get_sector_api_v1_map_sectors__sector_id__get
      parameters:
      - name: sector_id
        in: path
        required: true
        schema:
          type: string
          title: Sector Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SectorSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/gates:
    get:
      tags:
      - map
      summary: List Gates
      operationId: list_gates_api_v1_map_gates_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GateSummary'
                title: Response List Gates Api V1 Map Gates Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/zones:
    get:
      tags:
      - map
      summary: List Zones
      operationId: list_zones_api_v1_map_zones_get
      parameters:
      - name: sector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZoneSummary'
                title: Response List Zones Api V1 Map Zones Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/sectors/{sector_id}/zones:
    get:
      tags:
      - map
      summary: Get Sector Zones
      description: List all zones available in a specific sector.
      operationId: get_sector_zones_api_v1_map_sectors__sector_id__zones_get
      parameters:
      - name: sector_id
        in: path
        required: true
        schema:
          type: string
          title: Sector Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ZoneSummary'
                title: Response Get Sector Zones Api V1 Map Sectors  Sector Id  Zones
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/resources:
    get:
      tags:
      - map
      summary: List Resources
      description: List resource region entries. Filter by ware to find where a resource
        spawns.
      operationId: list_resources_api_v1_map_resources_get
      parameters:
      - name: ware
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by ware (ore, silicon, nividium, hydrogen, helium, methane,
            ice)
          title: Ware
        description: Filter by ware (ore, silicon, nividium, hydrogen, helium, methane,
          ice)
      - name: sector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by sector macro ID
          title: Sector Id
        description: Filter by sector macro ID
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceEntry'
                title: Response List Resources Api V1 Map Resources Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/sectors/{sector_id}/resources:
    get:
      tags:
      - map
      summary: Get Sector Resources
      description: List all resources available in a specific sector.
      operationId: get_sector_resources_api_v1_map_sectors__sector_id__resources_get
      parameters:
      - name: sector_id
        in: path
        required: true
        schema:
          type: string
          title: Sector Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResourceEntry'
                title: Response Get Sector Resources Api V1 Map Sectors  Sector Id  Resources
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/regions:
    get:
      tags:
      - map
      summary: List Regions
      operationId: list_regions_api_v1_map_regions_get
      parameters:
      - name: cluster_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
      - name: sector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RegionSummary'
                title: Response List Regions Api V1 Map Regions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/map/superhighways:
    get:
      tags:
      - map
      summary: List Superhighways
      operationId: list_superhighways_api_v1_map_superhighways_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SuperhighwaySummary'
                title: Response List Superhighways Api V1 Map Superhighways Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ClusterSummary:
      properties:
        cluster_id:
          type: string
          title: Cluster Id
        macro_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Macro Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        environment:
          anyOf:
          - type: string
          - type: 'null'
          title: Environment
        sun_class:
          anyOf:
          - type: string
          - type: 'null'
          title: Sun Class
        population_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Population Id
        max_population:
          anyOf:
          - type: integer
          - type: 'null'
          title: Max Population
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
      additionalProperties: false
      type: object
      required:
      - cluster_id
      title: ClusterSummary
...
    GateSummary:
      properties:
        from_zone_id:
          type: string
          title: From Zone Id
        to_zone_id:
          type: string
          title: To Zone Id
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
      additionalProperties: false
      type: object
      required:
      - from_zone_id
      - to_zone_id
      - kind
      title: GateSummary
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RegionSummary:
      properties:
        region_id:
          type: string
          title: Region Id
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        sector_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
      additionalProperties: false
      type: object
      required:
      - region_id
      title: RegionSummary
    ResourceEntry:
      properties:
        region_name:
          type: string
          title: Region Name
        sector_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
        ware:
          type: string
          title: Ware
        yield_level:
          type: string
          title: Yield Level
      additionalProperties: false
      type: object
      required:
      - region_name
      - sector_id
      - ware
      - yield_level
      title: ResourceEntry
    SectorSummary:
      properties:
        sector_id:
          type: string
          title: Sector Id
        cluster_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cluster Id
        macro_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Macro Id
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        sunlight:
          anyOf:
          - type: number
          - type: 'null'
          title: Sunlight
        economy:
          anyOf:
          - type: number
          - type: 'null'
          title: Economy
        security:
          anyOf:
          - type: number
          - type: 'null'
          title: Security
        tags:
          anyOf:
          - type: string
          - type: 'null'
          title: Tags
        access_licence:
          anyOf:
          - type: string
          - type: 'null'
          title: Access Licence
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
      additionalProperties: false
      type: object
      required:
      - sector_id
      - cluster_id
      - owner_faction
      title: SectorSummary
...
    SuperhighwaySummary:
      properties:
        from_zone_id:
          type: string
          title: From Zone Id
        to_zone_id:
          type: string
          title: To Zone Id
      additionalProperties: false
      type: object
      required:
      - from_zone_id
      - to_zone_id
      title: SuperhighwaySummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ZoneSummary:
      properties:
        zone_id:
          type: string
          title: Zone Id
        sector_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Sector Id
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
      additionalProperties: false
      type: object
      required:
      - zone_id
      - sector_id
      title: ZoneSummary